# openapi/requirements-api.yaml
openapi: 3.0.3
info:
  title: Kravhantering Requirements REST API
  version: 0.1.0
  description: >
    Static CI contract for the authenticated JSON REST API used by the
    requirements UI. This file is not served by the application at runtime.
servers:
  - url: http://localhost:3001
    description: Local prodlike CI server only.
...
security:
  - cookieSession: []
paths:
...
  /api/requirements:
    get:
      x-auth: session
      x-csrf: none
      x-cache: framework-default
      tags: [requirements]
      operationId: listRequirements
      summary: List requirements as JSON.
      parameters:
        - $ref: '#/components/parameters/Locale'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortDirection'
        - name: uniqueIdSearch
          in: query
          schema: { type: string, maxLength: 100 }
        - name: descriptionSearch
          in: query
          schema: { type: string, maxLength: 200 }
        - name: areaIds
          in: query
          schema:
            type: array
            maxItems: 100
            items: { type: integer, minimum: 1 }
          style: form
          explode: true
        - name: categoryIds
          in: query
          schema:
            type: array
            maxItems: 100
            items: { type: integer, minimum: 1 }
          style: form
          explode: true
        - name: typeIds
          in: query
          schema:
            type: array
            maxItems: 100
            items: { type: integer, minimum: 1 }
          style: form
          explode: true
        - name: qualityCharacteristicIds
          in: query
          schema:
            type: array
            maxItems: 100
            items: { type: integer, minimum: 1 }
          style: form
          explode: true
        - name: statuses
          in: query
          schema:
            type: array
            maxItems: 100
            items: { type: integer, minimum: 1 }
          style: form
          explode: true
        - name: normReferenceIds
          in: query
          schema:
            type: array
            maxItems: 100
            items: { type: integer, minimum: 1 }
          style: form
          explode: true
        - name: requirementPackageIds
          in: query
          schema:
            type: array
            maxItems: 100
            items: { type: integer, minimum: 1 }
          style: form
          explode: true
        - name: priorityLevelIds
          in: query
          schema:
            type: array
            maxItems: 100
            items: { type: integer, minimum: 1 }
          style: form
          explode: true
        - name: verifiable
          in: query
          schema:
            type: array
            maxItems: 100
            items:
              type: string
              enum: ['true', 'false']
          style: form
          explode: true
      responses:
        '200':
          description: Requirement list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RequirementListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      x-auth: session
      x-csrf: same-origin
      x-cache: framework-default
      tags: [requirements]
      operationId: createRequirement
      summary: Create a requirement draft.
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/XRequestedWith'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequirementCreateMutation' }
      responses:
        '201':
          description: Created requirement version.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RequirementCreateResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/requirements/{id}:
    parameters:
      - $ref: '#/components/parameters/RequirementId'
    get:
      x-auth: session
      x-csrf: none
      x-cache: framework-default
      tags: [requirements]
      operationId: getRequirement
      summary: Get a requirement with version history.
      responses:
        '200':
          description: Requirement detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RequirementDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      x-auth: session
      x-csrf: same-origin
      x-cache: framework-default
      tags: [requirements]
      operationId: editRequirement
      summary: Edit a requirement draft.
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/XRequestedWith'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequirementEditMutation' }
      responses:
        '200':
          description: Updated requirement version wrapper.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RequirementMutationResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

    delete:
      x-auth: session
      x-csrf: same-origin
      x-cache: framework-default
      tags: [requirements]
      operationId: archiveRequirement
      summary: Archive a requirement.
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/XRequestedWith'
      responses:
        '200':
          description: Archived requirement result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArchiveRequirementResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: kravhantering_session
...
  parameters:
    RequirementId:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 80
        pattern: '^.+$'
      description: Numeric database id or stable requirement unique id.
...
    Locale:
      name: locale
      in: query
      schema:
        type: string
        enum: [en, sv]
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        minLength: 1
        maxLength: 512
    SortBy:
      name: sortBy
      in: query
      schema:
        type: string
        enum:
          - uniqueId
          - description
          - area
          - category
          - type
          - qualityCharacteristic
          - priorityLevel
          - status
          - version
    SortDirection:
      name: sortDirection
      in: query
      schema:
        type: string
        enum: [asc, desc]
    Origin:
      name: Origin
      in: header
      required: true
      schema:
        type: string
        enum: [http://localhost:3001]
    XRequestedWith:
      name: X-Requested-With
      in: header
      required: true
      schema:
        type: string
        enum: [XMLHttpRequest]
  responses:
...
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Authentication required.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Forbidden.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Conflict:
      description: Conflict.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: true
      properties:
        code: { type: string }
        detail: { type: string }
        details: {}
        error: { type: string }
...
    RequirementListResponse:
      type: object
      required: [pagination, requirements]
      additionalProperties: false
      properties:
        pagination:
          type: object
          required: [count, hasMore, limit, nextCursor]
          additionalProperties: false
          properties:
            count: { type: integer, minimum: 0, maximum: 200 }
            hasMore: { type: boolean }
            limit: { type: integer, minimum: 1, maximum: 200 }
            nextCursor:
              type: string
              nullable: true
              maxLength: 512
        requirements:
          type: array
          minItems: 0
          maxItems: 200
          items: { $ref: '#/components/schemas/RequirementSummary' }
...
    RequirementSummary:
      type: object
      required: [id, uniqueId]
      additionalProperties: true
      properties:
        id: { type: integer }
        uniqueId: { type: string }
        isArchived: { type: boolean }
        version:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/RequirementVersion'
    RequirementDetail:
      type: object
      required: [id, uniqueId, permissions]
      additionalProperties: true
      properties:
        id: { type: integer }
        uniqueId: { type: string }
        isArchived: { type: boolean }
        version:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/RequirementVersion'
        versions:
          type: array
          minItems: 0
          maxItems: 100
          items: { $ref: '#/components/schemas/RequirementVersion' }
        permissions:
          $ref: '#/components/schemas/RequirementDetailPermissions'
    RequirementDetailPermissions:
      type: object
      required:
        - allowedTransitionStatusIds
        - canArchive
        - canDeleteDraft
        - canEdit
        - canManageSuggestions
        - canReactivate
        - canRestore
        - canViewHistory
      additionalProperties: false
      properties:
        allowedTransitionStatusIds:
          type: array
          minItems: 0
          maxItems: 20
          items: { type: integer }
        canArchive: { type: boolean }
        canDeleteDraft: { type: boolean }
        canEdit: { type: boolean }
        canManageSuggestions: { type: boolean }
        canReactivate: { type: boolean }
        canRestore: { type: boolean }
        canViewHistory: { type: boolean }
...
    RequirementMutationResult:
      type: object
      required: [version]
      additionalProperties: true
      properties:
        id: { type: integer }
        uniqueId: { type: string }
        version: { $ref: '#/components/schemas/RequirementVersion' }
    RequirementCreateResult:
      type: object
      required: [requirement, version]
      additionalProperties: true
      properties:
        requirement:
          type: object
          required: [id, uniqueId]
          additionalProperties: true
          properties:
            id: { type: integer }
            uniqueId: { type: string }
        version: { $ref: '#/components/schemas/RequirementVersion' }
...
    ArchiveRequirementResult:
      type: object
      required: [detail, id, ok, uniqueId]
      additionalProperties: true
      properties:
        detail:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/RequirementDetail'
        id: { type: integer, nullable: true }
        ok: { type: boolean }
        uniqueId: { type: string, nullable: true }
...
    RequirementVersion:
      type: object
      additionalProperties: true
      properties:
        id: { type: integer }
        versionNumber: { type: integer }
        description: { type: string, nullable: true }
        acceptanceCriteria: { type: string, nullable: true }
        verificationMethod: { type: string, nullable: true }
        verifiable: { type: boolean }
        status: { type: integer, nullable: true }
        statusNameSv: { type: string, nullable: true }
        statusNameEn: { type: string, nullable: true }
        revisionToken: { type: string }
    RequirementCreateMutation:
      type: object
      required: [areaId, description]
      additionalProperties: false
      properties:
        acceptanceCriteria: { type: string, maxLength: 4000 }
        areaId: { type: integer, minimum: 1 }
        categoryId: { type: integer, minimum: 1, nullable: true }
        description: { type: string, minLength: 1, maxLength: 4000 }
        normReferenceIds:
          type: array
          nullable: true
          maxItems: 200
          items: { type: integer, minimum: 1 }
        qualityCharacteristicId: { type: integer, minimum: 1, nullable: true }
        requirementPackageIds:
          type: array
          nullable: true
          maxItems: 200
          items: { type: integer, minimum: 1 }
        verifiable: { type: boolean }
        priorityLevelId: { type: integer, minimum: 1, nullable: true }
        typeId: { type: integer, minimum: 1, nullable: true }
        verificationMethod: { type: string, maxLength: 4000 }
    RequirementEditMutation:
      type: object
      required: [baseRevisionToken, baseVersionId, description]
      additionalProperties: false
      properties:
        acceptanceCriteria: { type: string, maxLength: 4000 }
        areaId: { type: integer, minimum: 1, nullable: true }
        baseRevisionToken:
          type: string
          format: uuid
          nullable: false
        baseVersionId:
          type: integer
          minimum: 1
          nullable: false
        categoryId: { type: integer, minimum: 1, nullable: true }
        description: { type: string, minLength: 1, maxLength: 4000 }
        normReferenceIds:
          type: array
          nullable: true
          maxItems: 200
          items: { type: integer, minimum: 1 }
        qualityCharacteristicId: { type: integer, minimum: 1, nullable: true }
        requirementPackageIds:
          type: array
          nullable: true
          maxItems: 200
          items: { type: integer, minimum: 1 }
        verifiable: { type: boolean }
        priorityLevelId: { type: integer, minimum: 1, nullable: true }
        typeId: { type: integer, minimum: 1, nullable: true }
        verificationMethod: { type: string, maxLength: 4000 }